heurestic solution for bad f clef pitchfinding
Before this, f clefs would always be found in relation to the f part of the clef, which is higher up than the c clef part which actually dictates the pitch.

Therefore, f clefs pitches are now  found by throwing away the left side of the image and using the c clef on the right instead.